Return held packages from package_list
package_list should return a list of all installed packages matching
a pattern, but the regexp excluded packages that were included-but-held.

In practice, this means pg_upgradecluster would fail to upgrade to a
version where the postgresql package was held. E.g. an upgrade from X to
Y where the postgresql-Y package was held would fail, saying the package
was missing.

To fix this, update the regexp to include both installed and held
packages.